chore: update bash and rust scripts
The bash_aliases script was updated to include a new binding for ~/.inputrc. In the run script, binding to ~/.inputrc was removed and sourcing of ~/.bashrc was added. The rust script was updated to source ~/.cargo/env instead of ~/.bashrc. Lastly, a fix was made to the sudo script to correctly use the $USER variable in the echo command.