chore: update bash_aliases and tmux.conf
This commit updates the bash_aliases file to add the "-a" option to the "l" alias command, and removes the "cd /mnt/c/Users/sett" statement. Additionally, this commit updates the tmux.conf file to install the tpm plugin, configure tmux to use the 256-color terminal, and bind some tmux keybindings.